requirements.txt
PyYAML
selenium
pytest
allure-pytest